+++ coldtype/text/reader.py
@@ -1,9 +1,7 @@
-from pathlib import Path
 from collections import OrderedDict
-from functools import partial, lru_cache
-from urllib.request import urlretrieve
+from functools import partial
 
-import unicodedata, math, os, re, tempfile
+import unicodedata, math
 
 from fontTools.misc.transform import Transform
 from fontTools.pens.transformPen import TransformPen
@@ -11,14 +9,10 @@ from fontTools.pens.transformPen import TransformPen
 from coldtype.color import normalize_color, rgb
 from coldtype.runon.path import P
 from coldtype.geometry import Rect
-
-from coldtype.osutil import on_linux, on_mac, on_windows
+from coldtype.text.font import Font, normalize_font_path, normalize_font_prefix, FontNotFoundException, ALL_FONT_DIRS, _prefixes
 
 from typing import Union
 
-from coldtype.fontgoggles.font import getOpener
-from coldtype.fontgoggles.font.baseFont import BaseFont
-from coldtype.fontgoggles.font.otfFont import OTFFont
 from coldtype.fontgoggles.misc.textInfo import TextInfo
 
 try:
@@ -28,8 +22,6 @@ try:
 except ImportError:
     BlackRendererFont = None
     pass
-
-BLACKRENDER_ALL = False
 
 class FittableMixin():
     def textContent(self):
@@ -56,263 +48,11 @@ class FittableMixin():
         return self
 
 
-_prefixes = [
-    ["¬", "~/Library/Fonts"],
-    ["", "/Library/Fonts"]
-]
-
-ALL_FONT_DIRS = []
-
-if on_mac():
-    ALL_FONT_DIRS = [
-        ".",
-        "/System/Library/Fonts",
-        "/Library/Fonts",
-        "~/Library/Fonts",
-    ]
-
-elif on_windows():
-    ALL_FONT_DIRS = [
-        ".",
-        "C:/Windows/Fonts",
-    ]
-
-    localappdata = os.environ.get("LOCALAPPDATA")
-    if localappdata:
-        ALL_FONT_DIRS.append(str(Path(localappdata) / "Microsoft/Windows/Fonts/"))
-
-elif on_linux():
-    ALL_FONT_DIRS = ["."]
-    # TODO what are the default linux font installation dirs?
-    pass
-
-FONT_FIND_DEPTH = 3
-
-class FontNotFoundException(Exception):
-    pass
-
-def normalize_font_prefix(path_string):
-    for prefix, expansion in _prefixes:
-        path_string = path_string.replace(prefix, expansion)
-    return Path(path_string).expanduser().resolve()
-
-def normalize_font_path(font, nonexist_ok=False):
-    global _prefixes
-    literal = normalize_font_prefix(str(font))
-    ufo = literal.suffix == ".ufo"
-    if nonexist_ok:
-        return str(literal)
-    if literal.exists() and (not literal.is_dir() or ufo):
-        return str(literal)
-    else:
-        raise FontNotFoundException(literal)
-
-FontCache = {}
-
-class Font():
-    # TODO support glyphs?
-    def __init__(self, path,
-        number=0,
-        cacheable=False,
-        suffix=None,
-        delete_tmp=False
-        ):
-        tmp = None
-        if isinstance(path, str) and path.startswith("http"):
-            url = Path(path)
-            sfx = url.suffix
-            if not sfx:
-                sfx = suffix
-            with tempfile.NamedTemporaryFile(prefix="coldtype_download_temp", suffix="."+sfx, delete=False) as tmp:
-                urlretrieve(path, tmp.name)
-                path = tmp.name
-                tmp = tmp
-        
-        self.path = Path(normalize_font_path(path))
-        numFonts, opener, getSortInfo = getOpener(self.path)
-        self.font:BaseFont = opener(self.path, number)
-        self.font.cocoa = False
-        self.cacheable = cacheable
-        self._loaded = False
-        self.load()
-
-        self._colr = self.font.ttFont.get("COLR")
-        self._colrv1 = (self._colr is not None
-            #and self._colr.version == 1
-            and BlackRendererFont is not None)
-
-        if self._colrv1 or BLACKRENDER_ALL:
-            self._brFont = BlackRendererFont(self.path, fontNumber=number)
-        else:
-            self._brFont = None
-
-        self._variations = self.font.ttFont.get("fvar")
-
-        if tmp and delete_tmp:
-            os.unlink(tmp.name)
-    
-    def load(self):
-        if self._loaded:
-            return self
-        else:
-            self.font.load(None)
-            self._loaded = True
-            return self
-    
-    def variations(self):
-        axes = {}
-        if self._variations:
-            fvar = self._variations
-            for axis in fvar.axes:
-                axes[axis.axisTag] = (axis.__dict__)
-        return axes
-    
-    @staticmethod
-    def Cacheable(path, suffix=None, delete_tmp=False, actual_path=None):
-        """use actual_path to override a key path (if the actual path is the result of a networked call)"""
-        if path not in FontCache:
-            FontCache[path] = Font(
-                actual_path if actual_path else path,
-                cacheable=True,
-                suffix=suffix,
-                delete_tmp=delete_tmp).load()
-        return FontCache[path]
-    
-    @staticmethod
-    def GDrive(id, suffix, delete=True):
-        dwnl = f"https://drive.google.com/uc?id={id}&export=download"
-        return Font.Cacheable(dwnl, suffix=suffix, delete_tmp=delete)
-    
-    @staticmethod
-    def GoogleFont(font_name, index=0) -> "Font":
-        import requests, zipfile, io
-
-        font_name_short = font_name.replace(" ", "")
-        font_cache_key = f"GoogleFont_{font_name_short}_{index}"
-        if font_cache_key in FontCache:
-            return FontCache[font_cache_key]
-
-        url = f"https://fonts.google.com/download?family={font_name}"
-        folder = Path(f"_GoogleFonts/{font_name_short}")
-        folder.mkdir(exist_ok=True, parents=True)
-
-        r = requests.get(url)
-        if not r.ok:
-            raise FontNotFoundException("GoogleFont URL did not resolve")
-        
-        z = zipfile.ZipFile(io.BytesIO(r.content))
-        z.extractall(folder)
-        
-        font_path = list(folder.glob("*.ttf"))[index]
-        return Font.Cacheable(font_cache_key, actual_path=font_path)
-    
-    def _ListDir(dir, regex, regex_dir, log=False, depth=0):
-        if dir.name in [".git", "venv"]:
-            return
-        
-        #print(dir.stem, depth, len(os.listdir(dir)))
-        results = []
-
-        for p in dir.iterdir():
-            if p.is_dir() and depth < FONT_FIND_DEPTH and p.suffix != ".ufo":
-                try:
-                    res = Font._ListDir(p, regex, regex_dir, log, depth=depth+1)
-                    if res:
-                        results.extend(res)
-                except PermissionError:
-                    pass
-            else:
-                if regex_dir and not re.search(regex_dir, str(p.parent)):
-                    continue
-                if re.search(regex, p.name):
-                    if p.suffix in [".otf", ".ttf", ".ttc", ".ufo"]:
-                        results.append(p)
-        
-        return results
-
-    @lru_cache()
-    def List(regex, regex_dir=None, log=False, font_dir=None, max_depth=FONT_FIND_DEPTH):
-        results = []
-        
-        font_dirs = ALL_FONT_DIRS
-        if font_dir is not None:
-            font_dirs = [font_dir]
-        
-        for dir in font_dirs:
-            dir = normalize_font_prefix(dir)
-            results.extend(Font._ListDir(Path(dir), regex, regex_dir, log, depth=0))
-        return sorted(results, key=lambda p: p.stem)
-
-    def Find(regex, regex_dir=None, index=0):
-        if isinstance(regex, Font):
-            return regex
-
-        if Path(normalize_font_prefix(regex)).expanduser().exists():
-            return Font.Cacheable(regex)
-        
-        found = Font.List(regex, regex_dir)
-        try:
-            return Font.Cacheable(found[index])
-        except Exception as e:
-            #print(">", e)
-            raise FontNotFoundException(regex)
-    
-    def RegisterDir(dir):
-        global ALL_FONT_DIRS
-        if dir not in ALL_FONT_DIRS:
-            ALL_FONT_DIRS.insert(0, dir)
-    
-    def Normalize(font, fallback=True):
-        if isinstance(font, Path):
-            font = str(font)
-        
-        if isinstance(font, str):
-            try:
-                _font = Font.Find(font)
-                _font.load() # necessary?
-                return _font
-            except FontNotFoundException as e:
-                if fallback:
-                    print("font not found:", font)
-                    return Font.RecursiveMono()
-                else:
-                    raise e
-        elif isinstance(font, Font):
-            return font
-        else: # it's a list of fonts
-            for f in font:
-                try:
-                    return Font.Normalize(f, fallback=False)
-                except FontNotFoundException:
-                    pass
-            if fallback:
-                return Font.RecursiveMono()
-            else:
-                raise FontNotFoundException()
-
-    @staticmethod
-    def ColdtypeObviously():
-        return Font.Cacheable(Path(__file__).parent.parent / "demo/ColdtypeObviously-VF.ttf")
-    
-    ColdObvi = ColdtypeObviously
-
-    @staticmethod
-    def MutatorSans():
-        return Font.Cacheable(Path(__file__).parent.parent / "demo/MutatorSans.ttf")
-    
-    MuSan = MutatorSans
-    
-    @staticmethod
-    def RecursiveMono():
-        return Font.Cacheable(Path(__file__).parent.parent / "demo/RecMono-CasualItalic.ttf")
-    
-    RecMono = RecursiveMono
-
 class Style():
     """
     Class for configuring font properties
 
-    **Keyword arguments**
+    #### Keyword arguments
 
     * `font`: can either be a `coldtype.text.Font` object, a `pathlib.Path`, or a plain string path
     * `font_size`: standard point-based font-size, expressed as integer
@@ -326,7 +66,7 @@ class Style():
     * `removeOverlaps` (aka `ro`): automatically use skia-pathops to remove overlaps from the glyphs (useful when using variable ttf fonts)
     * `lang`: set language directly, to access language-specific alternate characters/rules
 
-    **Shorthand kwargs**
+    #### Shorthand kwargs
 
     * `kp` for `kern_pairs` — a dict of glyphName->[left,right] values in font-space
     * `tl` for `trackingLimit`
@@ -338,51 +78,52 @@ class Style():
         _prefixes.append([prefix, str(expansion)])
 
     def __init__(self,
-            font:Union[Font, str]=None,
-            font_size:int=12,
-            tracking=0,
-            trackingMode=1,
-            kern_pairs=dict(),
-            space=None,
-            baselineShift=0,
-            xShift=None,
-            rotate=0,
-            reverse=False,
-            removeOverlap=False,
-            q2c=False,
-            lang=None,
-            narrower=None,
-            fallback=None,
-            palette=0,
-            capHeight=None,
-            ascender=None,
-            descender=None,
-            metrics="c",
-            data={},
-            layer=None,
-            liga=True,
-            kern=True,
-            fill=rgb(0, 0.5, 1),
-            stroke=None,
-            strokeWidth=0,
-            variations=dict(),
-            variationLimits=dict(),
-            trackingLimit=0,
-            scaleVariations=True,
-            rollVariations=False,
-            mods=None,
-            features=dict(),
-            increments=dict(),
-            varyFontSize=False,
-            preventHwid=False,
-            fitHeight=None,
-            meta=dict(),
-            no_shapes=False,
-            show_frames=False,
-            load_font=True, # should we attempt to load the font?
-            tag=None, # way to differentiate in __eq__
-            _stst=False,
-            **kwargs):
+        font:Union[Font, str]=None,
+        font_size:int=12,
+        tracking=0,
+        trackingMode=1,
+        kern_pairs=dict(),
+        space=None,
+        baselineShift=0,
+        xShift=None,
+        rotate=0,
+        reverse=False,
+        removeOverlap=False,
+        q2c=False,
+        lang=None,
+        narrower=None,
+        fallback=None,
+        palette=0,
+        capHeight=None,
+        ascender=None,
+        descender=None,
+        metrics="c",
+        data={},
+        layer=None,
+        liga=True,
+        kern=True,
+        fill=rgb(0, 0.5, 1),
+        stroke=None,
+        strokeWidth=0,
+        variations=dict(),
+        variationLimits=dict(),
+        trackingLimit=0,
+        scaleVariations=True,
+        rollVariations=False,
+        mods=None,
+        features=dict(),
+        increments=dict(),
+        varyFontSize=False,
+        preventHwid=False,
+        fitHeight=None,
+        meta=dict(),
+        no_shapes=False,
+        show_frames=False,
+        load_font=True, # should we attempt to load the font?
+        tag=None, # way to differentiate in __eq__
+        _stst=False,
+        **kwargs
+        ):
 
         self.input = locals()
         self.input["self"] = None