+++ CHANGELOG.md
@@ -200,12 +200,14 @@ as in the new truchet animation examples, and examples/animations/ec.py)
 ### Added
 - `P.gridlayer`, `Runon.mapvch`, `Runon.mapvrc`
 
-## [0.10.0] - 2022-??-??
+## [0.10.0] - 2023-01-15
 ### Added
 - `.up` as canonical form of `.ups`
 - Argument-less boolean operation methods for plural `P`
 - Automatic _-prefixed versions of all P methods, to get clojure-style "phrase"-commenting
 - `VERSIONS=` support for macro-like reversioning of a single animation
+- `Font.LibraryFind` and `Font.LibraryList` to search font registry (mac-only, windows is excruciatingly difficult)
+- `Font.names` to get style and family name of font via fontTools
 ### Changed
 - `Mondrian` -> `Scaffold`
 - `th` & `tv` (true-horizontal & true-vertical) have been renamed to potential confusion (since `th` could be taken to mean `true-height`) — these are now `tx` and `ty` respectively